Add files via upload
The code generates and prints the Fibonacci sequence, which is a sequence of numbers in which each number is the sum of the two preceding ones, starting from 0 and 1. The code defines a function fibonacci that takes a single argument n, which represents the number of terms to generate in the sequence.

The function first checks the input n for correctness. If n is less than or equal to 0, the function prints an error message. If n is equal to 1, the function returns 0, since the first term in the Fibonacci sequence is 0. If n is equal to 2, the function returns a list containing 0 and 1, which are the first two terms in the sequence.

For all other values of n, the function initializes a list fib_list with the first two terms of the sequence (0 and 1), and then uses a while loop to generate and append the remaining terms to the list. The loop continues until the length of the list fib_list is equal to the input n. The next term in the sequence is computed by adding the last two terms in the list fib_list.

Finally, the function returns the generated Fibonacci sequence. The code also includes a section where the user can specify the number of terms by

entering a value when prompted. The code then calls the fibonacci function with this input and prints the resulting sequence.

This code can be used to demonstrate the concept of the Fibonacci sequence and how it can be generated using a simple program in Python. It can also be used as a building block for more complex programs that require generating and manipulating the Fibonacci sequence, for example, in mathematical or computational applications.